forecast-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.27.0
pydantic==2.6.3
pydantic-settings==2.2.1
pandas==2.2.1
numpy==1.26.4
prophet==1.1.5
statsmodels==0.14.1
pmdarima==2.0.4
scikit-learn==1.4.1.post1
tensorflow==2.16.1
diskcache==5.6.3
tenacity==8.2.3
python-dotenv==1.0.1
opencensus-ext-azure==1.1.13
azure-storage-blob==12.19.1
pytest==8.0.2
pytest-asyncio==0.23.5
openpyxl==3.1.2
